fix sync on meta
