+++ common/common.go
@@ -6,8 +6,6 @@ import (
 	"github.com/multiversx/mx-chain-core-go/core"
 	"github.com/multiversx/mx-chain-core-go/core/check"
 	"github.com/multiversx/mx-chain-core-go/data"
-	"github.com/multiversx/mx-chain-go/storage"
-	"github.com/multiversx/mx-chain-vm-v1_2-go/ipc/marshaling"
 )
 
 // IsValidRelayedTxV3 returns true if the provided transaction is a valid transaction of type relayed v3
@@ -76,29 +74,4 @@ func VerifyProofAgainstHeader(proof data.HeaderProofHandler, header data.HeaderH
 	}
 
 	return nil
-}
-
-// GetHeader tries to get the header from pool first and if not found, searches for it through storer
-func GetHeader(
-	headerHash []byte,
-	headersPool HeadersPool,
-	headersStorer storage.Storer,
-	marshaller marshaling.Marshalizer,
-) (data.HeaderHandler, error) {
-	header, err := headersPool.GetHeaderByHash(headerHash)
-	if err == nil {
-		return header, nil
-	}
-
-	headerBytes, err := headersStorer.SearchFirst(headerHash)
-	if err != nil {
-		return nil, err
-	}
-
-	err = marshaller.Unmarshal(header, headerBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	return header, nil
 }
